working ADC

+++ transmitter/src/adc.h
@@ -1,7 +1,7 @@
 #ifndef ADC_H
 #define ADC_H
 
-#include <zephyr/drivers/adc.h>
+// #include <zephyr/drivers/adc.h>
 #include <zephyr/kernel.h>
 #include <zephyr/sys/printk.h>
 
@@ -14,7 +14,7 @@ int read_adc();
 
 int setup_adc();
 
-int tx_matched_filter(uint32_t *signal_buf, struct cached_sin_cos_t *cached_s_c);
+int tx_matched_filter(int16_t *signal_buf, struct cached_sin_cos_t *cached_s_c);
 
 int setup_sin_cos_cache(struct cached_sin_cos_t *cached_s_c);
